% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopping_functions.R
\name{split_bill}
\alias{split_bill}
\title{Split a Bill Among Multiple People: split_bill}
\usage{
split_bill(bill, people = 2)
}
\arguments{
\item{bill}{Numeric. The total bill amount to be split.}

\item{people}{Integer. The number of people sharing the bill (default is 2).}
}
\value{
A named numeric vector giving the amount each person owes, truncated
  to two decimal places. Names are "Person 1", "Person 2", etc.
}
\description{
Divides a total bill evenly among a specified number of people and returns the
amount each person should pay.
}
\examples{
# Split a $50 bill between 2 people
split_bill(50)

# Split a $120 bill among 4 people
split_bill(120, people = 4)

}
