% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_toss.R
\name{coin_toss}
\alias{coin_toss}
\title{Simulate a Coin Toss: coin_toss}
\usage{
coin_toss(...)
}
\value{
Character. Either "Heads" or "Tails".
}
\description{
Randomly returns either "Heads" or "Tails" to simulate a single coin flip.
}
\examples{
# Simulate a coin toss
coin_toss()

# Toss multiple times using replicate
replicate(5, coin_toss())

}
